+++ src/index.tsx
@@ -1,6 +1,6 @@
 import ReactDOM from 'react-dom/client';
-import './index.css';
 import App from './App';
+import './styles/index.css';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
