+++ src/styles/Player.css
